Move Refund Info to a tab
trans

FIX journal form showing fiscalprinter data for bank journals

Allow to use tax exclude prices in products and force to set "Tax-Included Price" in pos config

REF file structure splitting fp90iii.js

Move "Wrong tax configuration" check before creating receipt

Update library ePOS-Print and Fiscal Print API

IMP fiscal_epos_print writing fiscal receipt data to pos.order

show_popup instead of alert
