+++ fiscal_epos_print/__manifest__.py
@@ -17,13 +17,9 @@
     'website': 'https://github.com/OCA/l10n-italy',
     'depends': ['point_of_sale', 'pos_order_mgmt'],
     'data': [
-        'views/account_statement_view.xml',
+        'views/account.xml',
         'views/point_of_sale.xml',
         'views/assets.xml',
-    ],
-    'js': [
-        'static/src/js/fp90iii.js',
-        'static/lib/fiscalprint/fiscalprint.js'
     ],
     'qweb': [
         'static/src/xml/pos.xml'
