+++ fiscal_epos_print/views/point_of_sale.xml
@@ -24,13 +24,23 @@
             <field name="model">pos.order</field>
             <field name="inherit_id" ref="point_of_sale.view_pos_pos_form"/>
             <field name="arch" type="xml">
-                <xpath expr='//group[@name="order_fields"]' position="after">
-                    <group name="refund_field" string="Refund Info">
-                        <field name="refund_date"/>
-                        <field name="refund_report"/>
-                        <field name="refund_doc_num"/>
-                        <field name="refund_cash_fiscal_serial"/>
-                    </group>
+                <xpath expr='//notebook' position="inside">
+                    <page name="rt_data" string="RT data">
+                        <group>
+                            <group string="Receipt">
+                                <field name="fiscal_receipt_number"/>
+                                <field name="fiscal_receipt_amount"/>
+                                <field name="fiscal_receipt_date"/>
+                                <field name="fiscal_z_rep_number"/>
+                            </group>
+                            <group string="Refund Info">
+                                <field name="refund_date"/>
+                                <field name="refund_report"/>
+                                <field name="refund_doc_num"/>
+                                <field name="refund_cash_fiscal_serial"/>
+                            </group>
+                        </group>
+                    </page>
                 </xpath>
             </field>
         </record>
